feat(77): Added actions for tasks
