--- fortinet-fortirecon-brand-protection/task.py
@@ -1,29 +1,0 @@
-"""
-Copyright start
-MIT License
-Copyright (c) 2025 Fortinet Inc
-Copyright end
-"""
-
-from .make_rest_api_call import MakeRestApiCall
-from connectors.core.connector import get_logger
-from .task import create_task, update_task
-
-logger = get_logger("fortinet-fortirecon-easm")
-
-
-def create_task(config, params):
-    MK = MakeRestApiCall(config=config)
-    endpoint = "/security-orchestration/{org_id}/tasks"
-    payload = MK.build_payload(params)
-    print(payload)
-
-    response = MK.make_request(endpoint=endpoint, method="POST", params=payload)
-    return response
-
-
-def update_task(config, params):
-    MK = MakeRestApiCall(config=config)
-    endpoint = "/security-orchestration/{org_id}"+"/tasks/{0}".format(params.pop("task_id"))
-    response = MK.make_request(endpoint=endpoint, method="PATCH", params=params)
-    return response
